En funcionamiento obtención de imagenes desde backend

+++ src/main/java/com/pgp/dao/BovineDao.java
@@ -27,5 +27,5 @@ public interface BovineDao extends JpaRepository<Bovine,Long>{
 	List<BovineDto> bovinesAndIdentifiers ();
 	
 	@Query(value="SELECT b.ID, b.NAME AS Name, nvl(i.DIIO,'Sin arete') AS Diio, i.DATE_PLACEMENT AS datePlacement, b.DATE_BIRTH AS dateBirth, bo.NAME AS Mother , tb.NAME AS Type FROM HERNAN.BOVINES b LEFT JOIN HERNAN.IDENTIFIERS i on b.ID = i.BOVINE_ID JOIN HERNAN.BOVINES bo on b.MOTHER = bo.ID  JOIN HERNAN.TYPES_BOVINES tb on tb.ID=b.TYPE WHERE b.ID = ? AND b.DATE_SALE IS NULL AND (i.STATE='activo' or i.DIIO is NULL ) AND b.STATE = 'Vivo' AND  b.ID <> 0", nativeQuery=true)
-	BovineDto bovineAndIdentifier(int id);
+	BovineDto bovineAndIdentifier(Long id);
 }
